requirements.txt
python-telegram-bot==13.13.0
python-dotenv==0.21.0
pyTelegramBotAPI